Refactor geospatial.
