+++ lib/alchemist/geospatial.rb
@@ -1,0 +1,7 @@
+module Alchemist
+  class Measurement
+    def geospatial
+      Alchemist::Earth.new(self).geospatial
+    end
+  end
+end
