+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 $LOAD_PATH.unshift File.join(File.dirname(__FILE__), "..", "lib")
 
 require 'alchemist'
+require 'alchemist/geospatial'
 
 Alchemist.setup
